requirements.txt
bcrypt==4.3.0
cffi==1.17.1
colorama==0.4.6
cryptography==45.0.2
flexcache==0.3
flexparser==0.4
iniconfig==2.1.0
msvc_runtime==14.42.34433
packaging==25.0
paramiko==3.5.1
Pint==0.24.4
platformdirs==4.3.8
pluggy==1.6.0
pycparser==2.22
pyfrc==2025.1.0
PyNaCl==1.5.0
pynetconsole==2.0.4
pynetworktables==2021.0.0
pyntcore==2025.3.2.2
pytest==8.3.5
pytest-reraise==2.1.2
robotpy==2025.3.2.2
robotpy-cli==2024.0.0
robotpy-commands-v2==2025.3.2
robotpy-hal==2025.3.2.2
robotpy-halsim-gui==2025.3.2.2
robotpy-installer==2025.1.1
robotpy-native-ntcore==2025.3.2
robotpy-native-wpihal==2025.3.2
robotpy-native-wpilib==2025.3.2
robotpy-native-wpimath==2025.3.2
robotpy-native-wpinet==2025.3.2
robotpy-native-wpiutil==2025.3.2
robotpy-navx==2025.0.1.2
robotpy-wpilib-utilities==2025.1.0
robotpy-wpimath==2025.3.2.2
robotpy-wpinet==2025.3.2.2
robotpy-wpiutil==2025.3.2.2
ruff==0.11.11
tomli==2.2.1
tomlkit==0.13.2
typing_extensions==4.13.2
wpilib==2025.3.2.2
